CompressDirectory dramatic speed boost on large numbers of files due to the careful scan approach implementation. Multi volume compress support.
git-svn-id: https://sevenzipsharp.svn.codeplex.com/svn@25191 9b283d60-5439-405e-af05-b73fd8c4d996

+++ SevenZip/Common.cs
@@ -661,7 +661,7 @@ namespace SevenZip
         /// <summary>
         /// Gets or sets the size of the archive volume (0 for no volumes)
         /// </summary>
-        long VolumeSize { get; set; }
+        int VolumeSize { get; set; }
         /// <summary>
         /// Occurs when the next file is going to be packed.
         /// </summary>
